chore(feedback implementation): - move unit tests into models folder - change test descriptions - remove error key - correct typographical errors
